refactor(backend): extract common code to mixin classes
- Add MySQLTransactionMixin with isolation level management
- Add MySQLBackendMixin with non-I/O methods
- Update MySQLBackend/AsyncMySQLBackend to inherit from MySQLBackendMixin
- Update MySQLTransactionManager/AsyncMySQLTransactionManager to inherit from MySQLTransactionMixin
- Remove duplicate code from sync/async implementations

This reduces code duplication and improves maintainability by sharing
non-I/O logic between sync and async implementations.
